requirements.txt
requests-oauthlib
pandas
python-dotenv